test: move new distinct comment falsy values tests into comment falsy value test file
